Добавление отдельного контроллера и представления для сотрудников
