+++ WebStore/Views/Employees/EmployeeDetails.cshtml
@@ -1,0 +1,14 @@
+﻿@model WebStore.Models.Employee
+
+
+<div class="container">
+    <h1>Employee information id:@Model.Id</h1>
+
+    <p>Surname: @Model.Surname</p>
+    <p>Name: @Model.Name</p>
+    <p>Patronymic: @Model.Patronymic</p>
+    <p>Age: @Model.Age</p>
+
+    <hr />
+    <p><a href="/Home/Index">back to list</a></p>
+</div>
