+++ WebStore/Views/Employees/Index.cshtml
@@ -1,0 +1,7 @@
+﻿
+@{
+    ViewData["Title"] = "Index";
+}
+
+<h1>Index</h1>
+
