--- WebStore/Views/Home/EmployeeDetails.cshtml
@@ -1,18 +1,0 @@
-﻿@model WebStore.Models.Employee
-
-
-<div class="container">
-    <head>
-        <title>"Employee Details"</title>
-    </head>
-
-    <body>
-        <h1>Employee information id:@Model.Id</h1>
-
-        <p>Surname: @Model.Surname</p>
-        <p>Name: @Model.Name</p>
-        <p>Patronymic: @Model.Patronymic</p>
-        <p>Age: @Model.Age</p>
-        <p><a href="/Home/Index">back to list</a></p>
-    </body>
-</div>
